fix deleting
