+++ sites/svelte.dev/src/routes/(authed)/+layout.server.js
@@ -1,10 +1,7 @@
-import * as cookie from 'cookie';
 import * as session from '$lib/db/session';
 
 export async function load({ request }) {
-	const cookies = cookie.parse(request.headers.get('cookie') || '');
-
 	return {
-		user: await session.read(cookies.sid)
+		user: await session.from_cookie(request.headers.get('cookie'))
 	};
 }
